booking->point->hotel 4차

+++ booking/src/main/java/hotelbooking/domain/BookingRepository.java
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 @Repository
-public interface BookingRepository extends JpaRepository<Booking, Long> {
+public interface BookingRepository extends JpaRepository<Booking, Integer> {
     Booking findByBookId(Integer bookId);
 }
